Create Arithmetic Calculator
In this project I created a calculator that identifies spaces and parenthesis. The calculator is implemented by a tree in which each node is an operation. The nodes that are closer to the root have higher priority and will be executed prior to the nodes that are closer to the leaf nodes.